small cleanup and restructure of files
