+++ package.json
@@ -26,6 +26,7 @@
   "homepage": "https://github.com/BJK/backwax",
   "devDependencies": {
     "assets-webpack-plugin": "^2.2.0",
+    "babel": "^5.8.23",
     "babel-loader": "^5.3.2",
     "chai": "^3.2.0",
     "flux": "^2.0.3",
