+++ src/index.js
@@ -1,0 +1,3 @@
+export ActionPrototype from './lib/ActionPrototype'
+export ExtraStorage from './lib/extra_storage'
+export Dispatcher from './lib/ep-dispatcher'
