+++ test/index.js
@@ -1,6 +1,6 @@
 var libContext = require.context("./lib/", true, /Spec.js$/);
 libContext.keys().forEach(libContext);
 
-import FUHelpers from './support/ChaiAddons'
+import FUHelpers from '../src/testUtils/ChaiAddons'
 
 chai.use(FUHelpers)
